Add IndentNo, decimal qty system, Units admin, FBR ItemType grouping + 4 perms
Indent No on delivery challans
- IndentNo column on DeliveryChallan + DTO + Print DTO + Excel template dict
- ChallanForm / ChallanEditForm / POImportForm: optional Indent No field,
  unified header layout (Client + Site + Delivery on one row, PO Number
  + PO Date + Indent No on the next)
- Migration AddIndentNoToDeliveryChallan + matching merge fields
  ({{indentNo}} and {{#if indentNo}}) so print templates can render
  conditionally

Decimal quantity system + Units admin
- InvoiceItem.Quantity / DeliveryItem.Quantity flipped int -> decimal(18,4)
  so fractional UOMs (KG, Liter, Carat) round-trip up to 4 places
- Unit model gains AllowsDecimalQuantity flag; new UnitsController
  (GET /api/units, PUT /api/units/{id}) gated by config.units.manage
- Program.cs startup backfill: union of FBR /uom master list (32) +
  every UOM in use across ItemType / InvoiceItem / DeliveryItem,
  with decimal flag auto-set on KG / Liter / Gram / MT / Carat /
  Square Foot / Cubic Metre / etc.
- UnitRegistry.EnsureNamesAsync called from every save path
  (challan create / edit / import, bill update, item-type save) so
  any new UOM the operator types appears on the Units admin grid
  immediately
- PO parser + Excel importer switched int.Parse -> decimal.Parse so
  source files carrying "12.5 KG" round-trip cleanly
- Server-side validation rejects fractional qty for integer-only UOMs
  with HTTP 400 + clear actionable error
- Frontend: shared QuantityInput component reacts to picked UOM
  (step="0.0001" vs step="1"), formatQuantity helper for read-only
  displays (parseFloat(toFixed(4)) -- strips trailing zeros), wired
  into ChallanForm / ChallanEditForm / EditBillForm / POImportForm,
  and InvoiceForm read-only display
- Qty columns widened (130px on Challan forms, 120px in EditBill,
  9% in InvoiceForm) and right-aligned for clarity
- /units admin page + sidebar nav + App.jsx route
- ItemType save auto-creates Unit row (idempotent, race-protected)

FBR submission improvements + 3 new permissions
- invoices.fbr.exclude: per-bill Exclude/Include FBR toggle now has
  its own perm so a role can hold the toggle without full edit rights.
  Startup migration auto-grants it to roles holding
  invoices.manage.update.
- invoices.manage.update.itemtype.qty: strict superset of
  update.itemtype that ALSO permits Quantity edits (decimal validation
  applies). New PATCH /invoices/{id}/itemtypes-and-qty endpoint.
  EditBillForm gates Qty cell on lockQty; service flag
  allowQuantityEdit gates the qty side. Auto-granted to roles
  holding invoices.manage.update.
- FBR payload now groups items by ItemType (sum qty + value + retail)
  matching the Tax Invoice print exactly: 4 lines sharing
  "Hardware Items" -> 1 FBR item with productDescription
  = "Hardware Items", quantity = 4, value = sum. Falls back to
  per-line emission when any line lacks an ItemType.
- New invoices.fbr.preview permission + GET
  /api/fbr/{id}/preview-payload that builds the would-be JSON
  without sending. Skips pre-validate so incomplete bills can
  still be inspected.
- New FbrPreviewDialog: opened from the View FBR button on each bill
  row. Shows seller / buyer / document blocks, grouped items table
  with summed qty / value / sales tax / further tax / line total per
  group, totals footer with grand total, toggleable raw JSON, copy
  button.

+++ DTOs/DeliveryItemDto.cs
@@ -6,7 +6,11 @@
         public int? ItemTypeId { get; set; }
         public string ItemTypeName { get; set; } = "";
         public string Description { get; set; } = "";
-        public int Quantity { get; set; }
+        // Decimal so fractional UOMs (KG, Liter, etc.) round-trip correctly.
+        // The frontend formats with at most 4 decimal places (trailing zeros
+        // stripped) so e.g. 12.5 displays as "12.5", 0.0004 as "0.0004",
+        // 0.09 as "0.09" — never "0.0900".
+        public decimal Quantity { get; set; }
         public string Unit { get; set; } = "";
     }
 }
